pyproject.toml
[tool.poetry]
name = "aio-agents"
version = "0.0.1"
description = "Opinionated template for building llm agents"
authors = ["Oscar Bahamonde <oscar.bahamonde@pucp.edu.pe>"]
readme = "README.md"
packages = [{ include = "aio_agents", from = "." }]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
aiofauna = "^0.1.82"
boto3 = "^1.28.30"
openai = "^0.27.8"
numpy = "^1.25.2"
tiktoken = "^0.4.0"
jinja2 = "^3.1.2"
aioredis = "^2.0.1"
markdown-it-py = {extras = ["linkify"], version = "^3.0.0"}
bs4 = "^0.0.1"
lxml = "^4.9.3"
pypdf = "^3.15.2"
aiohttp-cors = "^0.7.0"
spacy = "^3.6.1"
uvicorn = "^0.23.2"
scipy = "^1.11.2"
pydub = "^0.25.1"
pandas = "^2.0.3"
celery = "^5.3.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
